+++ src/main/java/com/kurly/api/jpa/repository/BasketRepository.java
@@ -15,6 +15,6 @@ import org.springframework.stereotype.Repository;
  * 2024-04-26        hagjoon       최초 생성
  */
 @Repository
-public interface BasketRepository extends JpaRepository<Basket,Integer> {
+public interface BasketRepository extends JpaRepository<Basket,Long> {
 
 }
